[fix] notify the observer only once
The notify callback only needs to be triggered once, but was trigerred
twice from these methods:
- moveShapesInObject
- setObjectPose

Added a flag to disable the notification when moving shapes. The shape
poses are first updated without notification, then the object pose is updated
which sends a notification to the observer. Halves the processing time
needed to move the object shapes.